requirements.txt
llama-index
llama-index-llms-openai
openai
python-dotenv
